Use routes/blueprints instead of controllers; use 'ckanext.' for config
