+++ ckanext/ckanpackager/plugin.py
@@ -4,6 +4,7 @@
 # This file is part of ckanext-ckanpackager
 # Created by the Natural History Museum in London, UK
 
+from ckanext.ckanpackager import routes
 from ckanext.ckanpackager.lib.helpers import should_show_format_options
 from ckanext.ckanpackager.lib.utils import url_for_package_resource
 from ckanext.ckanpackager.logic import action, auth
@@ -16,11 +17,11 @@ config = {}
 class CkanPackagerPlugin(SingletonPlugin):
     ''' '''
     implements(interfaces.ITemplateHelpers, inherit=True)
-    implements(interfaces.IRoutes, inherit=True)
     implements(interfaces.IConfigurable)
     implements(interfaces.IConfigurer)
     implements(interfaces.IActions)
     implements(interfaces.IAuthFunctions)
+    implements(interfaces.IBlueprint, inherit=True)
 
     def configure(self, app_cfg):
         '''Implementation of IConfigurable.configure
@@ -28,17 +29,21 @@ class CkanPackagerPlugin(SingletonPlugin):
         :param app_cfg: config dictionary
 
         '''
-        config[u'url'] = app_cfg[u'ckanpackager.url']
-        config[u'secret'] = app_cfg[u'ckanpackager.secret']
-        config[u'allow_anon'] = app_cfg.get(u'ckanpackager.allow_anon', False)
+        config[u'url'] = app_cfg.get(u'ckanext.ckanpackager.url')
+        config[u'secret'] = app_cfg.get(u'ckanext.ckanpackager.secret')
+        config[u'allow_anon'] = app_cfg.get(u'ckanext.ckanpackager.allow_anon', False)
         # As per ckan.controllers.packager.resource_read - there is no API for getting
         #  this.
         config[u'datastore_api'] = u'%s/api/action' % app_cfg.get(u'ckan.site_url',
                                                                   u'').rstrip('/')
 
+    ## IBlueprint
+    def get_blueprint(self):
+        return routes.blueprints
+
     def get_helpers(self):
         '''Implementation of ITemplateHelpers:get_helpers
-        
+
         Provide a helper for create a download url from a resource id
 
 
@@ -48,27 +53,12 @@ class CkanPackagerPlugin(SingletonPlugin):
             u'should_show_format_options': should_show_format_options
             }
 
-    def before_map(self, map_route):
-        '''Implements Iroutes.before_map
-        
-        Add our custom download action
-
-        :param map_route: 
-
-        '''
-        map_route.connect(u'package_resource',
-                          '/dataset/{package_id}/resource/{resource_id}/package',
-                          controller=u'ckanext.ckanpackager.controllers.packager'
-                                     u':CkanPackagerController',
-                          action=u'package_resource')
-        return map_route
-
     def update_config(self, app_config):
         '''Implementation IConfigurer.update_config
-        
+
         Add our template directory
 
-        :param app_config: 
+        :param app_config:
 
         '''
         toolkit.add_template_directory(app_config, u'theme/templates')
